+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.0 [unreleased]
 
+- Added support for loading `bool` type (as `uint8` for now)
 - Return `NArray` object for rank-0 arrays
 
 ## 0.1.2